% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{fn}
\alias{fn}
\title{Stuff}
\usage{
fn(x)
}
\description{
Stuff here
}
\details{
\subsection{This should run}{\if{html}{\out{<div class="sourceCode r">}}\preformatted{fn(1)
}\if{html}{\out{</div>}}\preformatted{## [1] 1
}
}

\subsection{Failing section}{\if{html}{\out{<div class="sourceCode r">}}\preformatted{#stop("oh no")
}\if{html}{\out{</div>}}
}
}
