% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{fn}
\alias{fn}
\title{Stuff}
\usage{
fn()
}
\description{
Stuff here
}
\details{
\subsection{Failing section}{\if{html}{\out{<div class="sourceCode r">}}\preformatted{1 + 1
}\if{html}{\out{</div>}}\preformatted{## [1] 2
}
}
}
